Completed third unit test
Here I added I changed the initial GetMovieGoer() in my test class to have an actual viewing history, I then created a new GetNullMoviegoer(), this way I can test if they have a viewing history or not, if they do not have a viewing history then we will return empty, this is not breaking any changes from recent tests. I assume this is what you were looking for on the third test without talking it through.